// src/common/enums/codes/api-error-subcode.enum.ts
export enum ApiErrorSubCode {
  NOT_FOUND = 'NOT_FOUND',
  BAD_DATA = 'BAD_DATA',
  UNAUTHORIZED = 'UNAUTHORIZED',
  FORBIDDEN = 'FORBIDDEN',
  CONFLICT = 'CONFLICT',
  INTERNAL = 'INTERNAL',
  QB_NOT_CONNECTED = 'QB_NOT_CONNECTED',
  QB_TOKEN_EXPIRED = 'QB_TOKEN_EXPIRED',
  QB_API_ERROR = 'QB_API_ERROR',
}